+++ src/main/java/online/kingdomkeys/kingdomkeys/client/render/entity/TrainingDummyRenderer.java
@@ -7,16 +7,15 @@ import net.minecraft.client.renderer.entity.EntityRendererProvider;
 import net.minecraft.client.renderer.entity.LivingEntityRenderer;
 import net.minecraft.resources.ResourceLocation;
 import online.kingdomkeys.kingdomkeys.KingdomKeys;
-import online.kingdomkeys.kingdomkeys.client.model.entity.BarrelModel;
+import online.kingdomkeys.kingdomkeys.client.model.entity.TrainingDummyModel;
 import online.kingdomkeys.kingdomkeys.entity.TrainingDummyEntity;
-import online.kingdomkeys.kingdomkeys.entity.mob.GigaShadowEntity;
 
 public class TrainingDummyRenderer extends LivingEntityRenderer<TrainingDummyEntity, EntityModel<TrainingDummyEntity>> {
 
-    private static final ResourceLocation TEXTURE = ResourceLocation.fromNamespaceAndPath(KingdomKeys.MODID, "textures/entity/training_dummy.png");
+    private static final ResourceLocation TEXTURE = ResourceLocation.fromNamespaceAndPath(KingdomKeys.MODID, "textures/entity/mob/training_dummy.png");
 
     public TrainingDummyRenderer(EntityRendererProvider.Context context) {
-        super(context, new BarrelModel<>(context.bakeLayer(BarrelModel.LAYER_LOCATION)), 0);
+        super(context, new TrainingDummyModel<>(context.bakeLayer(TrainingDummyModel.LAYER_LOCATION)), 0);
     }
 
     @Override
@@ -25,9 +24,14 @@ public class TrainingDummyRenderer extends LivingEntityRenderer<TrainingDummyEnt
     }
 
     @Override
+    protected boolean shouldShowName(TrainingDummyEntity entity) {
+        return false;
+    }
+
+    @Override
     protected void scale(TrainingDummyEntity entitylivingbaseIn, PoseStack matrixStackIn, float partialTickTime) {
-        matrixStackIn.scale(2F, 3F, 2F);
-        matrixStackIn.translate(0F, 1.2F, 0F);
+        float scale = 0.7F;
+        matrixStackIn.scale(scale, scale, scale);
         super.scale(entitylivingbaseIn, matrixStackIn, partialTickTime);
     }